requirements.txt
uvicorn[standard]~=0.18
fastapi~=0.81
mysql-connector-python~=8.0
starlette~=0.19
pydantic~=1.10
python-multipart>=0.0.5
